All files moved to the 'functions' folder and pipeline updated to work
