+++ functions/dataChecker.py
